Terraform Agent/requirements.txt
openai>=1.6.0
langgraph>=0.0.26
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.23.2